Add option to not use connection_pool
Part of the Quick Docs documentation is the suggestion to not use a HTTP connection pool to improve failure handling when using a proxy.

This has previously been tested (in basho_bench tests) with a branch that use send_req_direct on a new spawned worker for every request.

This should add this as a permanent option.

Streamed requests still use the connection pool.  Streamed requests are deprecated, and so changing this isn't required.